+++ src/index.js
@@ -4,12 +4,12 @@ const findScriptBaseUrl = () => {
   const scripts = document.getElementsByTagName('script');
   for (let i = 0; i < scripts.length; i++) {
     const script = scripts[i];
-    if (script.src.includes('gw2aEmbeds.js')) {
-      return script.src.replace('/gw2aEmbeds.js', '');
+    if (script.src.includes('armory-embeds.js')) {
+      return script.src.replace('/armory-embeds.js', '');
     }
   }
 
-  throw new Error('Could not find gw2aEmbeds script.');
+  throw new Error('Could not find armory-embeds script.');
 };
 
 const init = () => {