+++ CHANGELOG.md
@@ -6,11 +6,16 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased][]
+### Changed
+- gw2aEmbeds to armory-embeds
+- Readme information
+
+### Fixed
+- Package.json not pointing to correct entrypoint
 
 ## [0.1.0][] - 2017-10-10
 ### Added
 - Initial commit
-
 
 [Unreleased]: https://github.com/madou/armory-embeds/compare/v0.1.0...HEAD
 [0.1.0]: https://github.com/madou/armory-embeds/tree/v0.1.0